// src/app/components/button/button-color.enum.ts
export enum ButtonColor {
  DEFAULT = 'button-color-default',
  PRIMARY = 'button-color-primary',
  SECONDARY = 'button-color-secondary',
  SUCCESS = 'button-color-success',
  INFO = 'button-color-info',
  WARNING = 'button-color-warning',
  DANGER = 'button-color-danger',
  LIGHT = 'button-color-light'
}